requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
sqlalchemy==2.0.22
pillow==10.0.1
python-dotenv==1.0.0
psutil==5.9.5